requirements.txt
gunicorn
fastcluster==1.2.6
Flask==2.0.2
Flask-Mail==0.9.1
Flask-REST==1.3
Flask-RESTful==0.3.9
matplotlib==3.5.1
numba==0.56.4
numpy==1.21.5
pandas==1.3.5
pingouin==0.5.1
python-decouple==3.6
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
statsmodels==0.13.2
Werkzeug==2.0.2
psycopg2-binary==2.9.5
Deprecated~=1.2.13
setuptools~=65.5.1